Make it possible to run tests under autodiscovery
