+++ tests/Cargo.toml
@@ -4,8 +4,6 @@ version = "0.0.0"
 authors = ["John-John Tedro <udoprog@tedro.se>"]
 edition = "2018"
 
-autotests = false
-
 [features]
 default = ["full", "tokio", "futures-executor"]
 full = ["rune-modules/full"]
@@ -20,7 +18,3 @@ rune-modules = { path = "../crates/rune-modules" }
 runestick = { path = "../crates/runestick" }
 lazy_static = "1.4.0"
 parking_lot = "0.11.1"
-
-[[test]]
-name = "test"
-path = "test.rs"
