+++ tests/tests/test_range.rs
@@ -1,3 +1,5 @@
+use rune_tests::*;
+
 #[test]
 fn test_range() {
     rune! { () =>
