+++ tests/tests/vm_generators.rs
@@ -1,3 +1,5 @@
+use rune_tests::*;
+
 #[test]
 fn test_simple_generator() {
     assert_eq! {
