+++ tests/tests/vm_lazy_and_or.rs
@@ -1,3 +1,5 @@
+use rune_tests::*;
+
 #[test]
 fn test_lazy_and_or() {
     assert_eq! {
