+++ tests/tests/vm_result.rs
@@ -1,3 +1,5 @@
+use rune_tests::*;
+
 #[test]
 fn test_result() {
     assert_eq! {
