+++ tests/tests/vm_test_instance_fns.rs
@@ -1,3 +1,5 @@
+use rune_tests::*;
+
 #[test]
 fn test_instance_kinds() {
     assert_eq! {
